alfa_bank_integration/requirements.txt
requests
fastapi
uvicorn
python-dotenv
